refactor(server): break the boot/reconciler partition loader cycle
Sharing the disk loader with the reconciler pointed a second edge
at boot, which already drives the reconciler, so the module-graph
guard failed on a boot <-> partition_reconciler cycle.

The loader was never a boot concern. Boot walks every owned
namespace at startup, the reconciler picks up one whose create
commits later, and both want the same hydrate-or-fence policy.
partition_helpers already claimed that shared role in its module
doc and already holds the counterpart build_partition_fresh, and
both callers already import it, so the edge points one way again.

Importing through boot::recovery would also satisfy the guard,
since granularity is the module file, but it leaves the reconciler
reaching into boot's internals for a routine it has equal claim to.

Pure move: the relocated lines are unchanged apart from ordering
the public entry point ahead of its helpers.
