+++ core/server/src/boot/recovery.rs
@@ -19,12 +19,8 @@
 
 use crate::boot::topology::{RosterCells, TcpTopology, build_cluster_roster};
 use crate::boot::wire_shell_handlers;
-use crate::partition_helpers::{
-    build_partition_fresh, configure_consumer_offsets, ensure_initial_segment,
-    open_partition_superblock,
-};
-use crate::segment_recovery::{RecoveredSegment, load_persisted_segments};
-use crate::server_error::{PartitionRecoveryRefusal, ServerError};
+use crate::partition_helpers::load_partition_or_fence;
+use crate::server_error::ServerError;
 use crate::session_manager::SessionManager;
 use crate::shell::{
     ServerMetadata, ServerShard, ShellHandlers, consensus_timers, repair_retry_ticks,
@@ -34,20 +30,14 @@ use consensus::{
     ClientTable, JoinMode, LocalPipeline, PipelineEntry, Sequencer, VsrConsensus, VsrRestore,
     VsrState,
 };
-use iggy_common::{
-    Aes256GcmEncryptor, EncryptorKind, IggyByteSize, IggyError, PartitionStats, TopicRuntimeOptions,
-};
+use iggy_common::{Aes256GcmEncryptor, EncryptorKind, IggyByteSize, TopicRuntimeOptions};
 use journal::Journal;
 use journal::prepare_journal::PrepareJournal;
 use journal::superblock::PingPongSuperblock;
 use message_bus::IggyMessageBus;
-use metadata::ReplicaIdentity;
 use metadata::impls::metadata::{IggySnapshot, StreamsFrontend};
 use metadata::stm::snapshot::Snapshot;
-use metadata::stm::stream::Partition;
-use partitions::{
-    IggyIndexWriter, IggyPartition, IggyPartitions, MessagesWriter, PartitionsConfig,
-};
+use partitions::{IggyPartitions, PartitionsConfig};
 use server_common::sharding::{IggyNamespace, PartitionLocation, ShardId};
 use shard::builder::IggyShardBuilder;
 use shard::metrics::ShardMetrics;
@@ -57,12 +47,10 @@ use shard::{
     ShardIdentity, TaggedSender,
 };
 use std::cell::RefCell;
-use std::path::PathBuf;
 use std::rc::Rc;
 use std::sync::Arc;
-use std::sync::atomic::Ordering;
 use std::time::Duration;
-use tracing::{error, info, warn};
+use tracing::{info, warn};
 
 #[allow(clippy::too_many_arguments, clippy::too_many_lines)]
 pub(in crate::boot) async fn build_shard_for_thread(
@@ -586,549 +574,6 @@ pub(in crate::boot) fn restore_metadata_consensus(
     }
 
     consensus
-}
-
-/// Recover this partition's persisted segment chain, stamping each segment
-/// with the topic's effective segment size (the per-topic value when the
-/// topic was created with one, else the shard-wide configured size).
-///
-/// The topic's effective `enforce_fsync` goes in for the same reason: it is
-/// what tells recovery whether a durable index entry the log cannot back is a
-/// benign torn index or previously durable data the log lost.
-async fn recover_partition_segments(
-    config: &ServerConfig,
-    namespace: IggyNamespace,
-    runtime_options: TopicRuntimeOptions,
-    stats: &PartitionStats,
-) -> Result<Vec<RecoveredSegment>, ServerError> {
-    let stream_id = namespace.stream_id();
-    let topic_id = namespace.topic_id();
-    let partition_id = namespace.partition_id();
-    let segment_size = runtime_options
-        .segment_size
-        .unwrap_or_else(|| IggyByteSize::from(iggy_common::DEFAULT_SEGMENT_SIZE));
-    let enforce_fsync = runtime_options
-        .enforce_fsync
-        .unwrap_or(iggy_common::DEFAULT_ENFORCE_FSYNC);
-    load_persisted_segments(
-        config,
-        stream_id,
-        topic_id,
-        partition_id,
-        segment_size,
-        enforce_fsync,
-        stats,
-    )
-    .await
-    .map_err(|source| {
-        error!(
-            stream_id,
-            topic_id,
-            partition_id,
-            error = %source,
-            "failed to load partition log during server bootstrap"
-        );
-        source
-    })
-}
-
-/// Recover an owned partition from its on-disk state.
-///
-/// Shared by boot and the reconciler so a partition this replica committed
-/// before a crash but re-learns only after restart (its WAL watermark trails
-/// the commit by one op) is hydrated from its segments like any other, not
-/// rebuilt over them. `Ok(None)` means the namespace was tombstoned here; the
-/// arms below say when. Errors are transient I/O, left to the caller.
-#[allow(clippy::too_many_arguments, clippy::too_many_lines)]
-pub async fn load_partition_or_fence(
-    config: &ServerConfig,
-    namespace: IggyNamespace,
-    partition_stats: Arc<PartitionStats>,
-    partition_metadata: &Partition,
-    topic_runtime: TopicRuntimeOptions,
-    cluster_id: u128,
-    self_replica_id: u8,
-    replica_count: u8,
-    bus: Rc<IggyMessageBus>,
-    partitions: &IggyPartitions<Rc<IggyMessageBus>, PingPongSuperblock>,
-) -> Result<Option<IggyPartition<Rc<IggyMessageBus>>>, ServerError> {
-    let stream_id = namespace.stream_id();
-    let topic_id = namespace.topic_id();
-    // Heap-pinned: the loader's and the rebuilder's futures side by side
-    // outgrow clippy's `large_futures` cap, and this runs once per partition.
-    match Box::pin(load_partition(
-        config,
-        namespace,
-        Arc::clone(&partition_stats),
-        partition_metadata,
-        topic_runtime,
-        cluster_id,
-        self_replica_id,
-        replica_count,
-        Rc::clone(&bus),
-    ))
-    .await
-    {
-        Ok(partition) => Ok(Some(partition)),
-        // ONE damaged local chain must not take the node down. The shapes
-        // this refuses are structural -- what a failed state-transfer
-        // quarantine leaves behind, or damage the recovery walk proved
-        // inside a segment. What follows depends on whether a peer can
-        // restore the data. With peers, the segment files are fenced
-        // aside (keeping the superblock so the group cannot re-enter
-        // view 0), the group is materialised fresh, and the ordinary
-        // rejoin path (repair, then state transfer on a refused floor)
-        // refills it. Single-replica, only a chain-shape refusal whose
-        // planned chain provably holds ZERO recoverable bytes still
-        // fences and rebuilds: nothing servable is at stake, so an empty
-        // rebuild hides no loss. The verdict variant alone is not that
-        // evidence -- a hole and an orphan empty segment both fire over
-        // fully populated chains -- which is why the gate reads the byte
-        // total the refusal carries. Every other refusal tombstones,
-        // leaving its files exactly where they are: a rebuilt empty
-        // partition answers polls exactly like a healthy empty one and
-        // hides the loss, while an unrouted namespace is a failure an
-        // operator can see.
-        Err(ServerError::PartitionRecoveryRefused { dir, reason, .. }) => {
-            let partition_dir = dir.to_string_lossy().into_owned();
-            let rebuild_for_rejoin = replica_count > 1
-                || matches!(
-                    reason,
-                    PartitionRecoveryRefusal::Hole {
-                        recoverable_bytes: 0,
-                        ..
-                    } | PartitionRecoveryRefusal::EmptyNonTailSegment {
-                        recoverable_bytes: 0,
-                        ..
-                    }
-                );
-            error!(
-                stream_id,
-                topic_id,
-                partition_id = partition_metadata.id,
-                partition_dir,
-                %reason,
-                "refusing the recovered segment chain"
-            );
-            // A pass-A refusal folded nothing into the stats (recovery
-            // counts only accepted chains), but the hydrate-reopen refusal
-            // arrives after a fully counted load, so clear them either way.
-            partition_stats.zero_out_all();
-            if !rebuild_for_rejoin {
-                // No quarantine here, mirroring the superblock arm below:
-                // a tombstone is only durable if its cause is. Fencing the
-                // chain aside would leave the next boot zero segments to
-                // walk, so it would re-seed from the surviving superblock,
-                // plant a fresh segment, and serve the partition empty
-                // with no refusal logged. Left at their real paths, the
-                // same files re-derive this verdict (and this log line)
-                // every boot, and the reconciler's tombstone gate keeps
-                // the namespace away from a fresh build, whose
-                // initial-segment open would truncate the oldest refused
-                // segment in place. The one refusal whose cause is NOT
-                // durable is `StorageSizeMismatch`: it fires from the
-                // reopen right after recovery truncated the same file, so
-                // the next boot re-walks the already-truncated bytes and,
-                // unless the length diverges again, accepts the chain
-                // instead of re-tombstoning -- acceptable for an
-                // assertion that the filesystem lied about a length.
-                // `%reason` repeated on purpose: this is the line an
-                // operator greps to enumerate dark partitions, so it has
-                // to carry the verdict on its own.
-                error!(
-                    stream_id,
-                    topic_id,
-                    partition_id = partition_metadata.id,
-                    partition_dir,
-                    %reason,
-                    "no peer replica holds this partition's data; leaving the refused \
-                     segment files in place and tombstoning it instead of serving it \
-                     empty"
-                );
-                partitions.tombstone(namespace);
-                return Ok(None);
-            }
-            match partitions::state_transfer::quarantine_segment_files(&partition_dir).await {
-                Ok(fenced_dir) => error!(
-                    stream_id,
-                    topic_id,
-                    partition_id = partition_metadata.id,
-                    fenced_dir,
-                    "quarantined the refused segment files; they are kept for inspection"
-                ),
-                Err(error) => {
-                    // NOT rebuilt: `build_partition_fresh` reaches
-                    // `ensure_initial_segment`, which opens segment 0 with
-                    // `file_exists = false` and TRUNCATES whatever the
-                    // failed quarantine left behind. The likeliest failures
-                    // (suffix cap exhausted, `create_dir_all`) move zero
-                    // files, so rebuilding would destroy the oldest segment
-                    // on the first attempt while the higher-offset survivors
-                    // keep refusing every boot -- a loop that never
-                    // terminates and eats the chain one segment at a time.
-                    // Tombstone instead: the namespace stays unmaterialised
-                    // and unrouted, the reconciler backs off, and an
-                    // operator still has every byte.
-                    error!(
-                        stream_id,
-                        topic_id,
-                        partition_id = partition_metadata.id,
-                        partition_dir,
-                        %error,
-                        "failed to quarantine the refused segment files; leaving this \
-                         partition tombstoned rather than rebuilding over them"
-                    );
-                    partitions.tombstone(namespace);
-                    return Ok(None);
-                }
-            }
-            Box::pin(build_partition_fresh(
-                config,
-                namespace,
-                partition_stats,
-                partition_metadata.created_revision,
-                topic_runtime,
-                cluster_id,
-                self_replica_id,
-                replica_count,
-                partition_metadata.created_view,
-                Rc::clone(&bus),
-            ))
-            .await
-            .map(Some)
-        }
-        // An untrustworthy superblock fences ONE group, not the node. The
-        // segment files stay exactly where they are -- unlike a refused
-        // chain, the data on disk is not the thing in doubt -- so there is
-        // nothing to quarantine and nothing to rebuild: rebuilding fresh
-        // would hand this replica a view-0 identity while a record it
-        // cannot read says otherwise. Tombstoned, the namespace stays
-        // unmaterialised and unrouted, the reconciler backs off, and an
-        // operator has every byte plus a message naming the directory.
-        Err(
-            error @ (ServerError::PartitionSuperblockIo { .. }
-            | ServerError::PartitionSuperblockVersionUnknown { .. }
-            | ServerError::PartitionSuperblockUnverifiable { .. }
-            | ServerError::PartitionSuperblockUndecodable { .. }
-            | ServerError::PartitionSuperblockIdentityMismatch { .. }),
-        ) => {
-            error!(
-                stream_id,
-                topic_id,
-                partition_id = partition_metadata.id,
-                %error,
-                "cannot trust this partition's durable consensus state; tombstoning the \
-                 partition instead of serving it"
-            );
-            partition_stats.zero_out_all();
-            partitions.tombstone(namespace);
-            Ok(None)
-        }
-        Err(error) => Err(error),
-    }
-}
-
-#[allow(clippy::too_many_arguments)]
-async fn load_partition(
-    config: &ServerConfig,
-    namespace: IggyNamespace,
-    stats: Arc<PartitionStats>,
-    partition_metadata: &Partition,
-    runtime_options: TopicRuntimeOptions,
-    cluster_id: u128,
-    self_replica_id: u8,
-    replica_count: u8,
-    bus: Rc<IggyMessageBus>,
-) -> Result<IggyPartition<Rc<IggyMessageBus>>, ServerError> {
-    let stream_id = namespace.stream_id();
-    let topic_id = namespace.topic_id();
-    let partition_id = namespace.partition_id();
-    // (view, log_view) come from the group's durable superblock when present;
-    // a present but unverifiable record already refused boot inside
-    // `open_partition_superblock`.
-    let partition_dir = config
-        .system
-        .get_partition_path(stream_id, topic_id, partition_id);
-    let (superblock, recovered_state) = open_partition_superblock(
-        &partition_dir,
-        ReplicaIdentity {
-            cluster: cluster_id,
-            replica_id: self_replica_id,
-            replica_count,
-        },
-    )
-    .await?;
-
-    // A recovered partition lost its journal state with the process: the
-    // partition journal is in-memory and segments carry no op numbers, so
-    // this replica cannot know the group's (op, commit) even when the
-    // superblock restored its view. In a cluster it boots as a
-    // quorum-invisible backup and probes for the current view
-    // (`RequestStartView`): the view's primary answers with a `StartView`,
-    // journal repair fills the rejoin window, and the commit floor settles
-    // at the serving peer's retention point. The probe re-broadcasts on its
-    // timeout, so it needs no live mesh at boot. Single-replica groups
-    // have no peer to ask and keep the plain init.
-    let join = if replica_count > 1 {
-        JoinMode::ProbeAsBackup {
-            await_state_transfer: false,
-        }
-    } else {
-        JoinMode::Init
-    };
-    // Request queue holds 2x the prepare depth (buffered requests drain as
-    // prepares commit); depth is the per-partition `[partition]` knob.
-    let prepare_queue_depth = config.partition.prepare_queue_depth;
-    let timers = consensus_timers(config);
-    let consensus = VsrConsensus::restored(
-        cluster_id,
-        self_replica_id,
-        replica_count,
-        namespace.inner(),
-        bus,
-        LocalPipeline::with_capacities(prepare_queue_depth, prepare_queue_depth * 2),
-        VsrRestore {
-            timers: &timers,
-            durable_view: recovered_state
-                .as_ref()
-                .map(|state| (state.view, state.log_view)),
-            view_fallback: None,
-            seed_view: None,
-            incarnation: None,
-            join,
-        },
-    );
-
-    // No prepare-timestamp floor is restored here: the partition consensus
-    // journal is non-durable today, so there is no persisted head to observe
-    // (unlike `restore_metadata_consensus`, which observes its restored head).
-    // When PartitionJournal becomes durable (the milestone named in the
-    // multi-shard wiring commit body), observe the restored head and the max
-    // recovered message timestamp here, or an NTP rewind across a restart could
-    // regress persisted `base_timestamp`.
-
-    let recovered_segments =
-        recover_partition_segments(config, namespace, runtime_options, &stats).await?;
-
-    let mut partition = IggyPartition::new(stats.clone(), consensus);
-    partition.set_runtime_options(runtime_options);
-    partition.set_superblock(superblock, recovered_state.as_ref());
-    // Recovered partitions honor the same config-surfaced ring ceilings as the
-    // fresh-create path (build_partition_fresh). Retention is already off for
-    // single-replica groups, so this only sizes the multi-replica ring.
-    partition.log.journal().inner.set_ring_caps(
-        config.partition.evicted_ring_capacity,
-        config.partition.evicted_ring_bytes_max.as_bytes_u64(),
-    );
-    partition.set_partition_dir(partition_dir.clone());
-    // Before the hydrate: the durable record is keyed by incarnation, so a
-    // `purge.gen` left behind by a previous life of this namespace reads 0.
-    partition.set_created_revision(partition_metadata.created_revision);
-    partition.hydrate_applied_purge_generation().await?;
-    hydrate_partition_log(
-        &mut partition,
-        &partition_dir,
-        stream_id,
-        topic_id,
-        partition_id,
-        recovered_segments,
-    )
-    .await?;
-
-    let sized_end = partition
-        .log
-        .segments()
-        .iter()
-        .filter(|segment| segment.size > IggyByteSize::default())
-        .map(|segment| segment.end_offset)
-        .max();
-    // An empty chain whose segment is named for a nonzero offset is the
-    // shape a state-transfer install (or its converge) plants at the group
-    // frontier after the origin GC'd everything: the file name carries the
-    // frontier, and re-minting offsets from 0 here would fork this
-    // replica's batch stamps from the rest of the group after a restart.
-    let empty_frontier = partition
-        .log
-        .segments()
-        .iter()
-        .map(|segment| segment.start_offset)
-        .max()
-        .filter(|&start| sized_end.is_none() && start > 0);
-    let current_offset = sized_end.or_else(|| empty_frontier.map(|start| start - 1));
-    partition.created_at = partition_metadata.created_at;
-    partition.recovered_durable_offset = sized_end;
-    // The OFFSET COUNTER is restored from that file name (above), but the
-    // `installed_frontier` CLAIM deliberately is not: the claim says "everything
-    // below me is represented here", and `converge_to_empty_after_failed_install`
-    // refuses to make it when staged segments were dropped -- yet a converge
-    // plants exactly the same empty `{frontier:020}.log` a legitimate empty
-    // install does, so boot provably cannot tell them apart. Re-deriving it here
-    // would hand the refused claim back: the repair floor stand-in would accept a
-    // commit floor over ops this replica holds zero bytes for, and the replica
-    // would pass the serve gate and offer that emptiness onward, making a peer
-    // unlink its own chain. Leaving it `None` costs one spurious full
-    // re-transfer on the legitimate empty-install restart; a false caught-up
-    // claim is not recoverable. A durable home for the frontier (the partition
-    // superblock already reserves a field) is what would settle it properly.
-    let counter = current_offset.unwrap_or(0);
-    partition.offset.store(counter, Ordering::Release);
-    partition.dirty_offset.store(counter, Ordering::Relaxed);
-    partition.should_increment_offset = current_offset.is_some();
-    // The durable frontier is a LOWER BOUND on top of what the segments proved:
-    // it is the only carrier left when the segments that named the frontier are
-    // gone (an all-GC'd origin's install, a crash inside the swap window), and
-    // taking the max means real recovered data always wins.
-    partition.restore_offset_frontier(recovered_state.as_ref());
-    let current_offset = partition.offset.load(Ordering::Acquire);
-
-    configure_consumer_offsets(&mut partition, config, namespace, current_offset)?;
-    ensure_initial_segment(&mut partition, config, stream_id, topic_id, partition_id).await?;
-
-    Ok(partition)
-}
-
-/// Reopen writers over a recovered segment chain.
-///
-/// Takes no `&ServerConfig`: every knob it needs is the partition's own
-/// resolved topic option now, which is the whole point of the per-topic move.
-async fn hydrate_partition_log(
-    partition: &mut IggyPartition<Rc<IggyMessageBus>>,
-    partition_dir: &str,
-    stream_id: usize,
-    topic_id: usize,
-    partition_id: usize,
-    recovered_segments: Vec<RecoveredSegment>,
-) -> Result<(), ServerError> {
-    // The partition's own resolved knobs, not the shard-wide config: a topic
-    // created with `enforce_fsync` or a per-topic `segment_size` must get them
-    // on the writers reopened over its recovered chain too, or a restart would
-    // silently drop back to the node defaults.
-    let runtime = partition.runtime_options();
-    let enforce_fsync = runtime
-        .enforce_fsync
-        .unwrap_or(iggy_common::DEFAULT_ENFORCE_FSYNC);
-    let segment_size = runtime
-        .segment_size
-        .unwrap_or_else(|| IggyByteSize::from(iggy_common::DEFAULT_SEGMENT_SIZE));
-    let preallocate_segments = runtime
-        .preallocate_segments
-        .unwrap_or(iggy_common::DEFAULT_PREALLOCATE_SEGMENTS);
-    for RecoveredSegment { segment, storage } in recovered_segments {
-        partition
-            .log
-            .add_persisted_segment(segment, storage, None, None);
-    }
-
-    if let Some(active_index) = partition.log.segments().len().checked_sub(1) {
-        let storage = &partition.log.storages()[active_index];
-        if let (
-            Some(messages_reader),
-            Some(index_reader),
-            Some(storage_messages_writer),
-            Some(storage_index_writer),
-        ) = (
-            storage.messages_reader.as_ref(),
-            storage.index_reader.as_ref(),
-            storage.messages_writer.as_ref(),
-            storage.index_writer.as_ref(),
-        ) {
-            let index_path = index_reader.path();
-            let start_offset = partition.log.segments()[active_index].start_offset;
-            // Share the storage's size counters: they are the write cursors.
-            // A private counter would let the append position diverge from the
-            // segment bookkeeping that index entries and poll bounds rely on.
-            let messages_size_counter = storage_messages_writer.size_counter();
-            let index_size_counter = storage_index_writer.size_counter();
-            partition.log.messages_writers_mut()[active_index] = Some(Rc::new(
-                MessagesWriter::new(
-                    &messages_reader.path(),
-                    messages_size_counter,
-                    enforce_fsync,
-                    true,
-                    preallocate_segments.then_some(segment_size),
-                )
-                .await
-                .map_err(|source| {
-                    error!(
-                        stream_id,
-                        topic_id,
-                        partition_id,
-                        path = %messages_reader.path(),
-                        error = %source,
-                        "failed to initialize persisted messages writer"
-                    );
-                    hydrate_reopen_error(
-                        source,
-                        partition_dir,
-                        stream_id,
-                        topic_id,
-                        partition_id,
-                        start_offset,
-                    )
-                })?,
-            ));
-            partition.log.index_writers_mut()[active_index] = Some(Rc::new(
-                IggyIndexWriter::new(&index_path, index_size_counter, enforce_fsync, true)
-                    .await
-                    .map_err(|source| {
-                        error!(
-                            stream_id,
-                            topic_id,
-                            partition_id,
-                            path = %index_path,
-                            error = %source,
-                            "failed to initialize persisted sparse index writer"
-                        );
-                        hydrate_reopen_error(
-                            source,
-                            partition_dir,
-                            stream_id,
-                            topic_id,
-                            partition_id,
-                            start_offset,
-                        )
-                    })?,
-            ));
-        }
-    }
-
-    Ok(())
-}
-
-/// Routes a hydrate-reopen writer failure. The seed-vs-stat divergence guard
-/// (`SegmentSizeMismatchAtOpen`) is a post-condition assertion on recovery's
-/// own truncation: pass C truncates every file to its recovered size before
-/// storage and writers reopen it, so the guard can only fire if the
-/// filesystem lied about a length or a change broke that truncate-then-open
-/// contract. Kept as defense-in-depth and routed as a structural refusal
-/// because a retried boot cannot help. Every other failure here (open, stat,
-/// sync) is transient I/O and stays node-fatal: a retried boot can still
-/// serve the partition, while fencing would quarantine healthy data (and at
-/// `replica_count = 1` tombstone the partition outright).
-fn hydrate_reopen_error(
-    source: IggyError,
-    partition_dir: &str,
-    stream_id: usize,
-    topic_id: usize,
-    partition_id: usize,
-    start_offset: u64,
-) -> ServerError {
-    match source {
-        IggyError::SegmentSizeMismatchAtOpen(on_disk_bytes, expected_bytes) => {
-            ServerError::PartitionRecoveryRefused {
-                dir: PathBuf::from(partition_dir),
-                stream_id,
-                topic_id,
-                partition_id,
-                reason: PartitionRecoveryRefusal::StorageSizeMismatch {
-                    start_offset,
-                    on_disk_bytes,
-                    expected_bytes,
-                },
-            }
-        }
-        transient => transient.into(),
-    }
 }
 
 #[cfg(test)]
